Refacto
Logic:
The FlashCard now respect more the memory logic.

ELM:
The project was created to learn ELM and Haskell. After couple of weeks
with haskell on one side and ELM on the other, I felt very frustrated
with ELM, not having typeclass and other important functional
programming paradigm. This was not satisfying from my perspective, and I
decide to switch the frontend to very simple components, in react.
This is not a SPA anymore. If you're really interested in ELM, you can
checkout the past commits with the ELM SPA, and see how it can done.

Haskell:
After couple of months developing in haskell, I'm refactoring this
application to be more maintainable, and be more powerfull. Also, this
is not a SPA anymore, and thus the haskell part is much heavier.
- The performance has been greatly improved by having a small cache
implemented. The HTML templates, the SQL Queries and the CSS are now
kept in memory.
- The Database layer is a custom one over the LibPQ bindings, with some
typeclass to map the LibPQ binary format of the types to the Haskell
one. It has been proven to me that it's enough for the use case so far.
- The authentication is now done in the haskell layer instead of the DB.
This has been very fun to developed in PG, but I now want to keep the
logic of the authentication in one place, explaining the move.
- The templating is done by the Mustache library, which is far enough
for my use case.
- The design has been revisited, with the Clay library.

Database Schema:
- Move word to flashcard table, with recto/verso/bucket definition

PostgreSQL:
- The email field is now protected with a domain (regex check)